Add tests for creating dictionaries

+++ src/main/scala/ot/geckopipe/domain/Gene.scala
@@ -2,7 +2,7 @@ package ot.geckopipe.domain
 
 case class Gene(chr: String,
                 gene_id: String,
-                tss: Int,
-                start: Int,
-                end: Int,
+                tss: Long,
+                start: Long,
+                end: Long,
                 biotype: String)
